fix: change SubDir's return attribute
